refactor: CarControllerTest에서 하던 우승자 테스트를 CarsTest에게 위임, 그리고 CarController.getWinners 메서드를 default에서 private 으로 변경
